SRP 원칙에 따른 클래스 책임 분배 작업
