+++ src/main/java/com/test/Json/JsonPATHConfig.java
@@ -1,0 +1,6 @@
+package com.test.Json;
+
+public class JsonPATHConfig {
+    public static final String DIR_PATH = "src/main/java/com/test/db/wiseSaying/";
+    public static final String sequencePATH = DIR_PATH + "/lastid.txt";
+}
