+++ src/main/java/com/test/db/WiseSaying/13.json
@@ -1,0 +1,5 @@
+{
+    "id": "13",
+    "author": "테스트",
+    "content": "123"
+}
