+++ src/main/java/com/test/db/WiseSaying/14.json
@@ -1,0 +1,5 @@
+{
+    "id": "14",
+    "author": "안녕하세요",
+    "content": "이창중"
+}
