+++ src/main/java/com/test/db/WiseSaying/data.json
@@ -42,4 +42,14 @@
     "id": "12",
     "author": "새로운 저장소 저장 테스트",
     "content": "이창중"
+},
+{
+    "id": "13",
+    "author": "테스트",
+    "content": "123"
+},
+{
+    "id": "14",
+    "author": "안녕하세요",
+    "content": "이창중"
 }]
