+++ src/main/java/com/test/db/WiseSaying/lastid.txt
@@ -1,1 +1,1 @@
-12
+14
